+++ packages/plugins/plugin-block/src/block-plugin.ts
@@ -1,31 +1,10 @@
-import type { Node } from '@milkdown/prose/model'
 import type { PluginSpec } from '@milkdown/prose/state'
 import { Plugin, PluginKey } from '@milkdown/prose/state'
 import { $ctx, $prose } from '@milkdown/utils'
 
 import { BlockService } from './block-service'
 import { withMeta } from './__internal__/with-meta'
-
-/// @internal
-export type FilterNodes = (node: Node) => boolean
-
-/// @internal
-export const defaultNodeFilter: FilterNodes = (node) => {
-  const { name } = node.type
-  if (name.startsWith('table') && name !== 'table')
-    return false
-
-  return true
-}
-
-/// A slice contains the block config.
-/// Possible properties:
-/// - `filterNodes`: A function to filter nodes that can be dragged.
-export const blockConfig = $ctx<{ filterNodes: FilterNodes }, 'blockConfig'>({ filterNodes: defaultNodeFilter }, 'blockConfig')
-
-withMeta(blockConfig, {
-  displayName: 'Ctx<blockConfig>',
-})
+import { blockConfig } from './block-config'
 
 /// @internal
 export const blockService = $ctx(new BlockService(), 'blockService')
