+++ packages/plugins/plugin-block/src/block-service.ts
@@ -6,8 +6,8 @@ import { NodeSelection } from '@milkdown/prose/state'
 import type { EditorView } from '@milkdown/prose/view'
 import throttle from 'lodash.throttle'
 
-import type { FilterNodes } from './block-plugin'
-import { blockConfig } from './block-plugin'
+import type { FilterNodes } from './block-config'
+import { blockConfig } from './block-config'
 import { removePossibleTable } from './__internal__/remove-possible-table'
 import { selectRootNodeByDom } from './__internal__/select-node-by-dom'
 import { serializeForClipboard } from './__internal__/serialize-for-clipboard'
